tightened up center display
