+++ ui/frontend/build_src/src/components/molecules/generatedImage/generatedImage.css.ts
@@ -1,20 +1,11 @@
-import { style } from "@vanilla-extract/css";
+import { style, globalStyle } from "@vanilla-extract/css";
 
 export const generatedImageMain = style({
   position: "relative",
 });
 
-// export const imageContain = style({
-//   width: "512px",
-//   height: "512px",
-//   backgroundColor: "black",
-//   display: "flex",
-//   justifyContent: "center",
-//   alignItems: "center",
-// });
-
-export const image = style({
-  // width: "512px",
-  // height: "512px",
-  // objectFit: "contain",
+globalStyle(`${generatedImageMain} img`, {
+  width: "100%",
+  height: "100%",
+  objectFit: "contain",
 });
