+++ ui/frontend/build_src/src/components/organisms/creationPanel/advancedSettings/propertySettings/index.tsx
@@ -4,11 +4,10 @@ import { useCreateUI } from "../../creationPanelUIStore";
 
 import {
   SettingItem,
+  MenuButton
 } from "../../../../../styles/shared.css";
 
-import {
-  MenuButton,
-} from "../advancedsettings.css";
+
 
 import { useTranslation } from "react-i18next";
