+++ ui/frontend/build_src/src/components/organisms/creationPanel/basicCreation/makeButton/makeButton.css.ts
@@ -2,29 +2,9 @@ import { style } from "@vanilla-extract/css";
 
 import { vars } from "../../../../../styles/theme/index.css";
 
-export const MakeButtonStyle = style({
+import { BrandedButton } from "../../../../../styles/shared.css";
+
+export const MakeButtonStyle = style([BrandedButton, {
   width: "100%",
-  backgroundColor: vars.colors.brand,
   fontSize: vars.fonts.sizes.Headline,
-  fontWeight: "bold",
-  color: vars.colors.text.normal,
-  padding: vars.spacing.small,
-  borderRadius: vars.trim.smallBorderRadius,
-
-  ":hover": {
-    backgroundColor: vars.colors.brandHover,
-  },
-
-  ":active": {
-    backgroundColor: vars.colors.brandActive,
-  },
-
-  ":disabled": {
-    backgroundColor: vars.colors.brandDimmed,
-    color: vars.colors.text.dimmed,
-  },
-
-  ":focus": {
-    outline: "none",
-  },
-});
+}]);
