+++ ui/frontend/build_src/src/components/organisms/creationPanel/imageModifiers/imageModifiers.css.ts
@@ -15,21 +15,21 @@ export const ImageModifierGrouping = style({
   marginTop: vars.spacing.medium,
 });
 
-export const MenuButton = style({
-  display: "block",
-  width: "100%",
-  textAlign: "left",
-  backgroundColor: "transparent",
-  color: vars.colors.text.normal,
-  border: "0 none",
-  cursor: "pointer",
-  padding: "0",
-  marginBottom: vars.spacing.medium,
-});
+// export const MenuButton = style({
+//   display: "block",
+//   width: "100%",
+//   textAlign: "left",
+//   backgroundColor: "transparent",
+//   color: vars.colors.text.normal,
+//   border: "0 none",
+//   cursor: "pointer",
+//   padding: "0",
+//   marginBottom: vars.spacing.medium,
+// });
 
-globalStyle(`${MenuButton}> h4`, {
-  color: "#e7ba71",
-});
+// globalStyle(`${MenuButton}> h4`, {
+//   color: "#e7ba71",
+// });
 
 export const ModifierListStyle = style({
   // marginBottom: vars.spacing.small,
