+++ ui/frontend/build_src/src/components/organisms/displayPanel/completedImages/index.tsx
@@ -26,26 +26,6 @@ export default function CompletedImages(
     clearDisplay();
   };
 
-  // const idDelim = "_batch";
-  //   if (completedQueries.length > 0) {
-  //     // map the completedImagesto a new array
-  //     // and then set the state
-  //     const temp = completedQueries
-  //       .map((query, index) => {
-  //         if (void 0 !== query) {
-  //           return query.output.map((data: ImageOutput, index: number) => {
-  //             return {
-  //               id: `${completedIds[index]}${idDelim}-${data.seed}-${index}`,
-  //               data: data.data,
-  //               info: { ...query.request, seed: data.seed },
-  //             };
-  //           });
-  //         }
-  //       })
-  //       .flat()
-  //       .reverse()
-  //       .filter((item) => void 0 !== item) as CompletedImagesType[]; // remove undefined items
-
   return (
     <div className={completedImagesMain}>
       {/* Adjust the dom do we dont do this check twice */}
