+++ ui/frontend/build_src/src/components/organisms/displayPanel/currentDisplay/currentDisplay.css.ts
@@ -1,0 +1,12 @@
+import { style } from '@vanilla-extract/css'
+
+// handles all 3
+export const currentDisplayMain = style({
+  display: 'flex',
+  flexDirection: 'column',
+  justifyContent: 'center',
+  alignItems: 'center',
+  height: '100%',
+  width: '100%',
+  padding: '0 0 0 0',
+})
