+++ ui/frontend/build_src/src/components/organisms/displayPanel/currentDisplay/index.tsx
@@ -1,13 +1,15 @@
-/* eslint-disable @typescript-eslint/naming-convention */
-
 import React, { useEffect, useState } from "react";
-import GeneratedImage from "../../../molecules/generatedImage";
-import { useImageCreate } from "../../../../stores/imageCreateStore";
+
 import { FetchingStates, useImageFetching } from "../../../../stores/imageFetchingStore";
-import { CompletedImagesType, useImageDisplay } from "../../../../stores/imageDisplayStore";
+import { useImageDisplay } from "../../../../stores/imageDisplayStore";
 
 import { API_URL } from "../../../../api";
-import { isGeneratorFunction } from "util/types";
+
+import {
+  currentDisplayMain,
+} from './currentDisplay.css';
+
+import ImageDisplay from "./imageDisplay";
 
 
 const IdleDisplay = () => {
@@ -65,64 +67,6 @@ const LoadingDisplay = () => {
   );
 };
 
-const ImageDisplay = ({ info, data }: CompletedImagesType) => {
-
-  const createFileName = () => {
-    const {
-      prompt,
-      seed,
-      num_inference_steps,
-      guidance_scale,
-      use_face_correction,
-      use_upscale,
-      width,
-      height,
-    } = info;
-
-    // Most important information is the prompt
-    let underscoreName = prompt.replace(/[^a-zA-Z0-9]/g, "_");
-    underscoreName = underscoreName.substring(0, 100);
-    // name and the top level metadata
-    let fileName = `${underscoreName}_Seed-${seed}_Steps-${num_inference_steps}_Guidance-${guidance_scale}`;
-    // Add the face correction and upscale
-    if (typeof use_face_correction == "string") {
-      fileName += `_FaceCorrection-${use_face_correction}`;
-    }
-    if (typeof use_upscale == "string") {
-      fileName += `_Upscale-${use_upscale}`;
-    }
-    // Add the width and height
-    fileName += `_${width}x${height}`;
-    // add the file extension
-    fileName += ".png";
-    // return fileName
-    return fileName;
-  };
-
-  const setRequestOption = useImageCreate((state) => state.setRequestOptions);
-
-  const _handleSave = () => {
-    const link = document.createElement("a");
-    link.download = createFileName();
-    link.href = data ?? "";
-    link.click();
-  };
-
-  const _handleUseAsInput = () => {
-    setRequestOption("init_image", data);
-  };
-
-  return (
-    <div className="imageDisplay">
-      <p> {info?.prompt}</p>
-      <GeneratedImage imageData={data} metadata={info}></GeneratedImage>
-      <div>
-        <button onClick={_handleSave}>Save</button>
-        <button onClick={_handleUseAsInput}>Use as Input</button>
-      </div>
-    </div>
-  );
-};
 
 export default function CurrentDisplay() {
 
@@ -130,7 +74,7 @@ export default function CurrentDisplay() {
   const currentImage = useImageDisplay((state) => state.currentImage);
 
   return (
-    <div className="current-display">
+    <div className={currentDisplayMain}>
 
       {status === FetchingStates.IDLE && <IdleDisplay />}
